Centered text in upgrade buttons
Carson: Can you please create a separate label for the current cost of an upgrade instead of coding in the initial cost and modifying the entire label with js? IDK if putting a tag in the js replacement text will work and I’m afraid to mess with it. Thanks!